平均访问时间属性修改

+++ perfect-commons/src/main/java/com/perfect/dto/account/NmsGroupReportDTO.java
@@ -45,7 +45,7 @@ public class NmsGroupReportDTO extends AccountIdDTO {
 
     private Double hopRate;    // 二跳率
 
-    private Long avgResTime;  // 平均访问时间
+    private Double avgResTime;  // 平均访问时间
 
     private Integer directTrans;    // 直接转化
 
@@ -197,11 +197,11 @@ public class NmsGroupReportDTO extends AccountIdDTO {
         this.hopRate = hopRate;
     }
 
-    public Long getAvgResTime() {
+    public Double getAvgResTime() {
         return avgResTime;
     }
 
-    public void setAvgResTime(Long avgResTime) {
+    public void setAvgResTime(Double avgResTime) {
         this.avgResTime = avgResTime;
     }
